Rename EmbedProviderBase to OEmbedProviderBase
Rename EmbedProviderBase to OEmbedProviderBase
Add an obsoleted  class EmbedProviderBase
Add todo's to all implementations
