+++ src/Umbraco.Core/Media/EmbedProviders/Flickr.cs
@@ -1,9 +1,10 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Net;
 using Umbraco.Cms.Core.Serialization;
 
 namespace Umbraco.Cms.Core.Media.EmbedProviders
 {
+    // TODO(V10) : change base class to OEmbedProviderBase
     public class Flickr : EmbedProviderBase
     {
         public override string ApiEndpoint => "http://www.flickr.com/services/oembed/";
