+++ src/Umbraco.Core/Media/EmbedProviders/Hulu.cs
@@ -1,8 +1,9 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using Umbraco.Cms.Core.Serialization;
 
 namespace Umbraco.Cms.Core.Media.EmbedProviders
 {
+    // TODO(V10) : change base class to OEmbedProviderBase
     public class Hulu : EmbedProviderBase
     {
         public override string ApiEndpoint => "http://www.hulu.com/api/oembed.json";
