+++ src/Umbraco.Core/Media/EmbedProviders/Vimeo.cs
@@ -1,8 +1,9 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using Umbraco.Cms.Core.Serialization;
 
 namespace Umbraco.Cms.Core.Media.EmbedProviders
 {
+    // TODO(V10) : change base class to OEmbedProviderBase
     public class Vimeo : EmbedProviderBase
     {
         public override string ApiEndpoint => "https://vimeo.com/api/oembed.xml";
